try to makefile

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: kkafmagh <kkafmagh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/07 15:50:29 by kkafmagh          #+#    #+#             */
-/*   Updated: 2024/11/08 16:11:14 by kkafmagh         ###   ########.fr       */
+/*   Updated: 2024/11/18 11:19:12 by kkafmagh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
